Implement AbstractScheduledService from TestManager
This will allow Fusion labs to install ProxyTestManagerModule, which will
automatically run TestManager. For interoperability with non-Fusion hosts,
Runnable is left in place, and delegates to the AbstractScheduledService
methods.

For safety, `runLock` guards `isRunning`, which is used by both `startUp` and
`run` to guard that there aren't two threads attempting to run the same
TestManager.

PiperOrigin-RevId: 791885136