Create TestManagerService for starting up the ProxyTestManager in LabServer
This leaves the existing TestManager interface in place and creates a wrapper
around it for starting it with ServiceManager. Runnable is left in place, and
delegates to the AbstractScheduledService methods. This is for hosts that
continue to use Runnable.

For safety, `runLock` guards `isRunning`, which is used by both `startUp` and
`run` to guard that there aren't two threads attempting to run the same
TestManager.

PiperOrigin-RevId: 792224388