Extract TestService-related Guice bindings to a new module.
This change moves the Guice providers for `FileResolver`,
`MessageSenderFinder`, and `ExecTestServiceImpl` from `LabServerModule` to
`TestServiceModule`, located within the MH Test Runner codebase.

This also migrates the provided values to use an annotated
`ListeningExecutorService` rather than an unannotated one. This more narrowly
scopes the thread pool to only the things provided by this module.

PiperOrigin-RevId: 793911557